+++ Assets/SEE/Game/Evolution/Animators/MoveScaleShakeAnimator.cs
@@ -18,6 +18,7 @@
 //USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using SEE.Layout;
+using SEE.Utils;
 using System;
 using UnityEngine;
