+++ Assets/SEE/Game/GameNodeMover.cs
@@ -1,5 +1,4 @@
 ﻿using SEE.DataModel.DG;
-using SEE.Game.Evolution;
 using SEE.GO;
 using SEE.Utils;
 using UnityEngine;
